requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.10.2
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.7
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
argparse==1.2.1
decorator==3.4.0
itsdangerous==0.24
pbr==0.10.0
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.12
webassets==0.10.1
wsgiref==0.1.2
